os/bluestore: Move compression type enum and corresponding stuff to Compressor.h
