+++ src/compressor/Compressor.h
@@ -15,18 +15,19 @@
 #ifndef CEPH_COMPRESSOR_H
 #define CEPH_COMPRESSOR_H
 
-#include "include/int_types.h"
-#include "include/Context.h"
+#include <string>
+#include "include/memory.h"
+#include "include/buffer.h"
 
 class Compressor;
 typedef shared_ptr<Compressor> CompressorRef;
 
 class Compressor {
-  string type;
+  std::string type;
  public:
-  Compressor(string t) : type(t) {}
+  Compressor(std::string t) : type(t) {}
   virtual ~Compressor() {}
-  const string& get_type() const {
+  const std::string& get_type() const {
     return type;
   }
   virtual int compress(const bufferlist &in, bufferlist &out) = 0;
@@ -35,7 +36,32 @@ class Compressor {
   // alignment with decode methods
   virtual int decompress(bufferlist::iterator &p, size_t compressed_len, bufferlist &out) = 0;
 
-  static CompressorRef create(CephContext *cct, const string &type);
+  static CompressorRef create(CephContext *cct, const std::string &type);
+
+  enum CompressionAlgorithm {
+    COMP_ALG_NONE = 0,
+    COMP_ALG_SNAPPY = 1,
+    COMP_ALG_ZLIB = 2,
+  };
+
+  static const char * get_comp_alg_name(int a) {
+    switch (a) {
+    case COMP_ALG_NONE: return "none";
+    case COMP_ALG_SNAPPY: return "snappy";
+    case COMP_ALG_ZLIB: return "zlib";
+    default: return "???";
+    }
+  }
+
+  static int get_comp_alg_type(const std::string &s) {
+    if (s == "snappy")
+      return COMP_ALG_SNAPPY;
+    if (s == "zlib")
+      return COMP_ALG_ZLIB;
+
+    return COMP_ALG_NONE;
+  }
+
 };
 
 #endif
